Switch from WebIDL arrays to FrozenArray<>s in the IDLs.
Arrays have not existed in the WebIDL spec since 2015, so the Gamepad
spec (as well as the accompanying Gamepad Extensions one) were providing
invalid IDL definitions.

Switch to FrozenArray<T> and clarify in prose that some attributes should be
cached by the user agent until new values have to be returned.

Fixes #28.